requirements.txt
altair==5.2.0
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dlib==19.24.1
easyocr==1.7.1
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.1
fonttools==4.50.0
fsspec==2024.3.0
gitdb==4.0.11
GitPython==3.1.42
greenlet==3.0.3
idna==3.6
imageio==2.34.0
importlib-metadata==7.1.0
importlib-resources==6.3.1
ipykernel==6.29.3
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-client==8.6.1
jupyter-core==5.7.2
keras==2.15.0
kiwisolver==1.4.5
lazy-loader==0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
packaging==23.2
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pkgutil-resolve-name==1.3.10
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.1
pyclipper==1.3.0.post5
pydeck==0.8.1b0
pygments==2.17.2
pyparsing==3.1.2
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.28
stack-data==0.6.3
streamlit==1.37.1
sympy==1.12
tenacity==8.2.3
tensorflow==2.15.0
tifffile==2023.7.10
tk==0.1.0
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tornado==6.4
traitlets==5.14.2
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
zipp==3.18.1